shape flexibility
