+++ ocpmodels/trainers/base_trainer.py
@@ -18,7 +18,6 @@ import numpy.typing as npt
 import torch
 import torch.nn as nn
 import torch.optim as optim
-import torch_geometric
 import yaml
 from torch.nn.parallel.distributed import DistributedDataParallel
 from torch.utils.data import DataLoader
@@ -798,144 +797,6 @@ class BaseTrainer(ABC):
             self.optimizer.step()
         if self.ema:
             self.ema.update()
-
-    # Takes in a new data source and generates predictions on it.
-    @torch.no_grad()
-    def predict(
-        self,
-        data_loader,
-        per_image: bool = True,
-        results_file: Optional[str] = None,
-        disable_tqdm: bool = False,
-    ):
-        ensure_fitted(self._unwrapped_model, warn=True)
-
-        if distutils.is_master() and not disable_tqdm:
-            logging.info("Predicting on test.")
-        assert isinstance(
-            data_loader,
-            (
-                torch.utils.data.dataloader.DataLoader,
-                torch_geometric.data.Batch,
-            ),
-        )
-        rank = distutils.get_rank()
-
-        if isinstance(data_loader, torch_geometric.data.Batch):
-            data_loader = [data_loader]
-
-        self.model.eval()
-        if self.ema is not None:
-            self.ema.store()
-            self.ema.copy_to()
-
-        predictions = defaultdict(list)
-
-        for i, batch in tqdm(
-            enumerate(data_loader),
-            total=len(data_loader),
-            position=rank,
-            desc="device {}".format(rank),
-            disable=disable_tqdm,
-        ):
-
-            with torch.cuda.amp.autocast(enabled=self.scaler is not None):
-                out = self._forward(batch)
-
-            for target_key in self.config["outputs"]:
-                pred = out[target_key]
-                if self.normalizers.get(target_key, False):
-                    pred = self.normalizers[target_key].denorm(pred)
-
-                if per_image:
-                    ### Save outputs in desired precision, default float16
-                    if (
-                        self.config["outputs"][target_key].get(
-                            "prediction_dtype", "float16"
-                        )
-                        == "float32"
-                        or self.config["task"].get(
-                            "prediction_dtype", "float16"
-                        )
-                        == "float32"
-                        or self.config["task"].get("dataset", "lmdb")
-                        == "oc22_lmdb"
-                    ):
-                        dtype = torch.float32
-                    else:
-                        dtype = torch.float16
-
-                    pred = pred.cpu().detach().to(dtype)
-                    ### Split predictions into per-image predictions
-                    if (
-                        self.config["outputs"][target_key].get(
-                            "level", "system"
-                        )
-                        == "atom"
-                    ):
-                        batch_natoms = batch.natoms
-                        batch_fixed = batch.fixed
-                        per_image_pred = torch.split(
-                            pred, batch_natoms.tolist()
-                        )
-
-                        ### Save out only free atom, EvalAI does not need fixed atoms
-                        _per_image_fixed = torch.split(
-                            batch_fixed, batch_natoms.tolist()
-                        )
-                        _per_image_free_preds = [
-                            _pred[(fixed == 0).tolist()].numpy()
-                            for _pred, fixed in zip(
-                                per_image_pred, _per_image_fixed
-                            )
-                        ]
-                        _chunk_idx = np.array(
-                            [
-                                free_pred.shape[0]
-                                for free_pred in _per_image_free_preds
-                            ]
-                        )
-                        per_image_pred = _per_image_free_preds
-                    ### Assumes system level properties are of the same dimension
-                    else:
-                        per_image_pred = pred.numpy()
-                        _chunk_idx = None
-
-                    predictions[f"{target_key}"].extend(per_image_pred)
-                    ### Backwards compatibility, retain 'chunk_idx' for forces.
-                    if _chunk_idx is not None:
-                        if target_key == "forces":
-                            predictions["chunk_idx"].extend(_chunk_idx)
-                        else:
-                            predictions[f"{target_key}_chunk_idx"].extend(
-                                _chunk_idx
-                            )
-                else:
-                    predictions[f"{target_key}"] = pred.detach()
-
-            if not per_image:
-                return predictions
-
-            ### Get unique system identifiers
-            sids = batch.sid.tolist()
-            ## Support naming structure for OC20 S2EF
-            if "fid" in batch:
-                fids = batch.fid.tolist()
-                systemids = [f"{sid}_{fid}" for sid, fid in zip(sids, fids)]
-            else:
-                systemids = [f"{sid}" for sid in sids]
-
-            predictions["ids"].extend(systemids)
-
-        for key in predictions:
-            predictions[key] = np.array(predictions[key])
-
-        self.save_results(predictions, results_file)
-
-        if self.ema:
-            self.ema.restore()
-
-        return predictions
 
     def save_results(
         self, predictions, results_file: Optional[str], keys=None
